search_image/requirements.txt
fastapi
uvicorn
torch
torchvision
Pillow
scikit-learn
numpy
python-multipart
requests
